+++ src/popup/pages/SignTransaction/ZkAppChainView/index.js
@@ -128,8 +128,9 @@ const ZkAppChainView = ({ notifyParams, onRemoveNotify }) => {
       };
       await extSaveLocal(NET_WORK_CONFIG_V2, config);
       dispatch(updateCurrentNode(config.currentNode));
-        dispatch(updateCustomNodeList(config.customNodeList));
-
+      dispatch(updateCustomNodeList(config.customNodeList));
+      dispatch(updateShouldRequest(true));
+      dispatch(updateStakingRefresh(true));
       await extSaveLocal(NET_WORK_CHANGE_FLAG, true);
       sendNetworkChangeMsg(config.currentNode);
       clearLocalCache();